Add idiomatic rust iterator without cloning
Previously, there were two iterators that users could use. The first is
non-idiomatic iterator that matches RocksDB semantics, is flexible, and
performant. The second is more idiomatic, less flexible, and less
performant. Specifically, it would clone all keys and values while
iterating.

This commit adds a third iterator, that's the same as the idiomatic
iterator, but it doesn't clone keys and values, it returns references
to them. The existing idiomatic iterator was also re-written to wrap
this new iterator.